requirements.txt
opencv_python
onnxruntime-gpu
# onnxruntime
loguru